SyncContextUtil: fix initialization in Edit Mode
As described modesttree#9, using Unity3dAsyncAwaitUtil in
Editor scripts causes an NPE, because `SyncContextUtil.Install` is never called
to initialize global variables `SyncContextUtil.UnitySynchronizationContext` and
`SyncContextUtil.UnityThreadId`.

Fix this by adding the `[InitializeOnLoadMethod()]` attribute to the
`SyncContextUtil.Install` method. Note that the existing
`[RuntimeInitializeOnLoadMethod]` attribute is still needed in order for
`SyncContextUtil.Install` to be correctly invoked in Play Mode.

Fixes modesttree#9